+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - New #155: Add `NativeTransport` that used native `file_get_contents()` and `file_put_contents()` functions.
 - New #162: Add `ChecklistTask`, `Checklist`, `InputChecklistTask`, `InputChecklist`, `ChecklistTasksDone`,
   `ChecklistTasksAdded` and `DirectMessagePriceChanged` types.
-- New #162: Add `checklist`, `checklistTasksDone` and `checklistTasksAdded` fields to `Message` type.
+- New #162: Add `checklist`, `checklistTasksDone`, `checklistTasksAdded` and `directMessagePriceChanged` fields to
+  `Message` type.
 - New #162: Add `checklist` field to `ExternalReplyInfo` type.
 - New #162: Add `sendChecklist`, `editMessageChecklist` and `getMyStarBalance` methods.
 - New #162: Add `nextTransferDate` field to `OwnedGiftUnique` type.